Fixed bug and put settings in view files
