+++ webviz_subsurface/plugins/_well_analysis/_views/__init__.py
@@ -1,8 +1,2 @@
-from ._settings import (
-    ControlPressureOptions,
-    ControlSettings,
-    OverviewFilter,
-    OverviewPlotSettings,
-)
-from ._well_control import ControlView
-from ._well_overview import OverviewView
+from ._well_control import ControlPressureOptions, ControlSettings, ControlView
+from ._well_overview import OverviewFilter, OverviewPlotSettings, OverviewView
